chore: Organize client files
